requirements.txt
# Telegram Bot
python-telegram-bot>=20.0

# Steam API
steampy>=0.14.0

# Async support
aiohttp>=3.8.0

# Deep comparison
deepdiff>=6.0.0

# Database
aiosqlite>=0.19.0

# Logging
python-json-logger>=2.0.0

# Utils
python-dotenv>=1.0.0
